WIP: Storage mets manager
