+++ src/DigitalPreservation/Storage.Repository.Common/Mets/MetsManager.cs
@@ -1,6 +1,4 @@
-﻿using Amazon.S3;
-using Amazon.S3.Model;
-using Amazon.S3.Util;
+﻿using Amazon.Runtime.Internal.Util;
 using DigitalPreservation.Common.Model;
 using DigitalPreservation.Common.Model.Mets;
 using DigitalPreservation.Common.Model.Results;
@@ -10,18 +8,12 @@ using DigitalPreservation.Utils;
 using DigitalPreservation.XmlGen.Extensions;
 using DigitalPreservation.XmlGen.Mets;
 using DigitalPreservation.XmlGen.Premis.V3;
-using System.Net;
-using System.Text;
-using System.Xml;
-using System.Xml.Serialization;
-using Checksum = DigitalPreservation.Utils.Checksum;
-using File = System.IO.File;
 
 namespace Storage.Repository.Common.Mets;
 
 public class MetsManager(
-    IMetsParser metsParser, 
-    IAmazonS3 s3Client) : IMetsManager
+    IMetsParser metsParser,
+    IMetsStorage metsStorage) : IMetsManager
 {
     private const string PhysIdPrefix = "PHYS_";
     private const string FileIdPrefix = "FILE_";
@@ -171,152 +163,12 @@ public class MetsManager(
     
     public async Task<Result> WriteMets(FullMets fullMets)
     {
-        // TODO - re-use serializer? re-use XmlSerializerNamespaces GetNamespaces()?
-        
-        var serializer = new XmlSerializer(typeof(DigitalPreservation.XmlGen.Mets.Mets));
-        var sb = new StringBuilder();
-        var writer = XmlWriter.Create(sb, new XmlWriterSettings
-        {
-            OmitXmlDeclaration = true,
-            NamespaceHandling = NamespaceHandling.OmitDuplicates,
-            Encoding = Encoding.UTF8,
-            Indent = true,
-        });
-        serializer.Serialize(writer, fullMets.Mets, GetNamespaces());
-        writer.Close();
-        var xml = sb.ToString();
-
-        switch (fullMets.Uri.Scheme)
-        {
-            case "file":
-                await File.WriteAllTextAsync(fullMets.Uri.LocalPath, xml);
-                return Result.Ok();
-            
-            case "s3":
-                var awsUri = new AmazonS3Uri(fullMets.Uri);
-                var req = new PutObjectRequest
-                {
-                    BucketName = awsUri.Bucket,
-                    Key = awsUri.Key,
-                    ContentType = "application/xml",
-                    ContentBody = xml,
-                    ChecksumAlgorithm = ChecksumAlgorithm.SHA256
-                };
-                if (fullMets.ETag != null)
-                {
-                    req.IfMatch = fullMets.ETag;
-                }
-                var resp = await s3Client.PutObjectAsync(req);
-                if (resp.HttpStatusCode is HttpStatusCode.Created or HttpStatusCode.OK)
-                {
-                    return Result.Ok();
-                }
-                if (resp.HttpStatusCode is HttpStatusCode.PreconditionFailed)
-                {
-                    return Result.Fail(ErrorCodes.PreconditionFailed, "Supplied ETag did not match METS");
-                }
-                return Result.Fail(ErrorCodes.BadRequest, "AWS returned HTTP Status " + resp.HttpStatusCode + " when writing METS");
-            
-            default:
-                return Result.Fail(ErrorCodes.BadRequest, fullMets.Uri.Scheme + " not supported");
-        }
+        return await metsStorage.WriteMets(fullMets);
     }
 
     public async Task<Result<FullMets>> GetFullMets(Uri metsLocation, string? eTagToMatch)
     {
-        DigitalPreservation.XmlGen.Mets.Mets? mets = null; 
-        var fileLocResult = await metsParser.GetRootAndFile(metsLocation);
-        var (_, file) = fileLocResult.Value;
-        if (file is null)
-        {
-            return Result.FailNotNull<FullMets>(ErrorCodes.NotFound, "No METS file in " + metsLocation);
-        }
-        
-        string? returnedETag;
-
-        switch (file.Scheme)
-        {
-            case "s3":
-                var s3Uri = new AmazonS3Uri(file);
-                var gor = new GetObjectRequest
-                {
-                    BucketName = s3Uri.Bucket,
-                    Key = s3Uri.Key,
-                };
-                if (eTagToMatch is not null)
-                {
-                    gor.EtagToMatch = eTagToMatch;
-                }
-                try
-                {
-                    var resp = await s3Client.GetObjectAsync(gor);
-                    if (resp.HttpStatusCode == HttpStatusCode.OK)
-                    {
-                        var serializer = new XmlSerializer(typeof(DigitalPreservation.XmlGen.Mets.Mets));
-                        using var reader = XmlReader.Create(resp.ResponseStream);
-                        mets = (DigitalPreservation.XmlGen.Mets.Mets)serializer.Deserialize(reader)!;
-                    }
-                    if (resp.HttpStatusCode is HttpStatusCode.PreconditionFailed)
-                    {
-                        return Result.FailNotNull<FullMets>(ErrorCodes.PreconditionFailed, "Supplied ETag did not match METS");
-                    }
-                    returnedETag = resp.ETag;
-                }
-                catch (Exception e)
-                {
-                    return Result.FailNotNull<FullMets>(ErrorCodes.UnknownError, e.Message);
-                }
-                break;
-            
-            case "file":
-                var fi = new FileInfo(file.LocalPath);
-                try
-                {
-                    returnedETag = Checksum.Sha256FromFile(fi);
-                    if (eTagToMatch is not null && returnedETag != eTagToMatch)
-                    {
-                        return Result.FailNotNull<FullMets>(
-                            ErrorCodes.PreconditionFailed, "Supplied ETag did not match METS");
-                    }
-                    var serializer = new XmlSerializer(typeof(DigitalPreservation.XmlGen.Mets.Mets));
-                    using var reader = XmlReader.Create(file.LocalPath);
-                    mets = (DigitalPreservation.XmlGen.Mets.Mets)serializer.Deserialize(reader)!;
-                }
-                catch (Exception e)
-                {
-                    return Result.FailNotNull<FullMets>(ErrorCodes.UnknownError, e.Message);
-                }
-                break;
-            
-            default:
-                return Result.FailNotNull<FullMets>(ErrorCodes.BadRequest, file.Scheme + " not supported");
-        }
-
-
-        string? agentName = null;
-        if (mets?.MetsHdr?.Agent is not null && mets.MetsHdr.Agent.Count > 0)
-        {
-            agentName = mets.MetsHdr.Agent[0].Name;
-        }
-
-        if (agentName != IMetsManager.MetsCreatorAgent)
-        {
-            return Result.FailNotNull<FullMets>(ErrorCodes.BadRequest, "METS file was not created by " + IMetsManager.MetsCreatorAgent);
-        }
-
-        if (mets != null)
-        {
-            var fullMetal = new FullMets
-            {
-                Mets = mets,
-                Uri = file,
-                ETag = returnedETag
-            };
-            return Result.OkNotNull(fullMetal);
-        }
-
-        return Result.FailNotNull<FullMets>(
-            ErrorCodes.UnknownError, "Unable to read METS");
+       return await metsStorage.GetFullMets(metsLocation, eTagToMatch);
     }
 
     public async Task<Result> HandleSingleFileUpload(Uri workingRoot, WorkingFile workingFile, string depositETag)
@@ -346,7 +198,7 @@ public class MetsManager(
             {
                 await WriteMets(fullMets);
                 return Result.Ok();
-            }
+            } 
 
             return editMetsResult;
         }
@@ -797,16 +649,7 @@ public class MetsManager(
         return mets;
     }
     
-    private static XmlSerializerNamespaces GetNamespaces()
-    {
-        var ns = new XmlSerializerNamespaces();
-        ns.Add("mets", "http://www.loc.gov/METS/");
-        ns.Add("mods", "http://www.loc.gov/mods/v3");
-        ns.Add("premis", "http://www.loc.gov/premis/v3");
-        ns.Add("xlink", "http://www.w3.org/1999/xlink");
-        ns.Add("xsi", "http://www.w3.org/2001/XMLSchema-instance");
-        return ns;
-    }
+
     
     
     private static AmdSecType GetAmdSecType(FileFormatMetadata premisFile, string admId, string techId, string? digiprovId = null, VirusScanMetadata? virusScanMetadata = null, ExifMetadata? exifMetadata = null)
