+++ src/DigitalPreservation/XmlGen.Tests/MetsManagerTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Text.Json;
 using Amazon.S3;
 using DigitalPreservation.Common.Model;
+using DigitalPreservation.Common.Model.Mets;
 using DigitalPreservation.Common.Model.Transit;
 using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
@@ -24,10 +25,11 @@ public class MetsManagerTests
         var factory = serviceProvider.GetService<ILoggerFactory>();
         var parserLogger = factory!.CreateLogger<MetsParser>();
         var s3Client = new Mock<IAmazonS3>().Object;
+        
         parser = new MetsParser(s3Client, parserLogger);
-        metsManager = new MetsManager(parser, s3Client);
+        var metsStorage = new MetsStorage(s3Client, parser);
+        metsManager = new MetsManager(parser, metsStorage);
     }
-
 
     [Fact]
     public async Task Can_Create_Empty_Mets()
