+++ src/DigitalPreservation/XmlGen.Tests/ModsTests.cs
@@ -24,7 +24,8 @@ public class ModsTests
         var parserLogger = factory!.CreateLogger<MetsParser>();
         var s3Client = new Mock<IAmazonS3>().Object;
         parser = new MetsParser(s3Client, parserLogger);
-        metsManager = new MetsManager(parser, s3Client);
+        var metsStorage = new MetsStorage(s3Client, parser);
+        metsManager = new MetsManager(parser, metsStorage);
     }
     
     [Fact]
